Dialog boxes added
I have created the required dialog box classes inside main.java and they
are working. So I think we can remove window.xml

The submit button is only partially visible in the screen. We have to
relocate it.

Besides there are many other things which I need to add. But a primitive
model of the app is almost complete.