Edited
A few minute bugs fixed and some changes made!

What needs to be done:
a) the third member check box is clickable any time. It should only be
after the entries of two team members is complete.
b) On unchecking the check box the pink subheadings of name 3 and entry
number 3 are visible which is awkward and needs to be fixed.
c) An exit and reset button needs to be added.